requirements.txt
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
et-xmlfile==1.0.1
gcloud==0.17.0
googleapis-common-protos==1.51.0
httplib2==0.17.1
idna==2.7
isort==4.3.21
jdcal==1.4.1
jws==0.1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==3.0.0
openpyxl==3.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pylint==2.4.4
Pyrebase==3.0.27
python-jwt==2.0.1
python-dotenv==0.13.0
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.24.3
wrapt==1.11.2
